% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/add_files-method.R
\docType{methods}
\name{add_filePaths}
\alias{add_filePaths}
\alias{add_filePaths,skyline-method}
\title{add_filePaths}
\usage{
add_filePaths(object, filePaths)

\S4method{add_filePaths}{skyline}(object, filePaths)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{filePaths}{a vector of valid file (`.raw`, `.mzML` or `.mxXML`) locations for analysis}
}
\description{
add_filePaths

Add the file paths of all raw data to be analysed to a \code{skyline} object. The method writes to
the object in the global environment
}
