% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/add_files-method.R
\name{add_files}
\alias{add_files}
\alias{add_files,skyline-method}
\title{add_files}
\usage{
add_files(object, files)

\S4method{add_files}{skyline}(object, files)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{files}{a vector of valid file (\code{.raw}, \code{.mzML} or \code{.mxXML}) locations for analysis}
}
\description{
Add the file paths of all raw data to be analysed to a \code{skyline} object. The method writes to
the object in the global environment
}
