% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/add_path-method.R
\docType{methods}
\name{add_path}
\alias{add_path}
\alias{add_path,skyline-method}
\title{add_path}
\usage{
add_path(object, path)

\S4method{add_path}{skyline}(object, path)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{path}{a character string a an existing directory to be used for temporary file writing}
}
\description{
add_path

Add a `path` location which will be used as a temporay save to write-out `Skyline` result files
}
