% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/add_phenoData-method.R
\docType{methods}
\name{add_phenoData}
\alias{add_phenoData}
\alias{add_phenoData,skyline-method}
\title{add_phenoData}
\usage{
add_phenoData(object, phenoData)

\S4method{add_phenoData}{skyline}(object, phenoData)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{phenoData}{a runinfo \code{data.frame}}
}
\description{
add_phenoData

Add the runinfo \code{data.frame} to a \code{skyline} object. The method writes to
the object in the global environment
}
