% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/add_transitions.R
\docType{methods}
\name{add_transitions}
\alias{add_transitions}
\alias{add_transitions,skyline-method}
\title{add_transitions}
\usage{
add_transitions(object, transitions)

\S4method{add_transitions}{skyline}(object, transitions)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{transitions}{a \code{data.frame} of transition information. The \code{data.frame} must contain the following columns;
PrecursorName, PrecursorRT, PrecursorMz, ProductMz,PrecursorCharge and ProductCharge}
}
\description{
add_transitions

Add a transition \code{data.frame} to a \code{skyline} object. The method writes to
the object in the global environment
}
