% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/addFiles-method.R
\docType{methods}
\name{addFiles}
\alias{addFiles}
\alias{addFiles,Skyline-method}
\title{addFiles}
\usage{
addFiles(object, filepath)

\S4method{addFiles}{Skyline}(object, filepath)
}
\arguments{
\item{object}{a \code{Skyline} object}

\item{filepath}{a valid filepath to directory of \code{.raw} files}
}
\description{
addFiles

Add the file paths of all raw data to be analysed to a \code{Skyline} object. The method writes to
the object in the global environment
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

