% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/addInfo-method.R
\docType{methods}
\name{addInfo}
\alias{addInfo}
\alias{addInfo,skyline-method}
\title{addInfo}
\usage{
addInfo(object, runinfo)

\S4method{addInfo}{skyline}(object, runinfo)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{a}{runinfo \code{data.frame}}
}
\description{
addInfo

Add the runinfo \code{data.frame} to a \code{skyline} object. The method writes to
the object in the global environment
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

