% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/addTransitions-method.R
\docType{methods}
\name{addTransitions}
\alias{addTransitions}
\alias{addTransitions,Skyline-method}
\title{addTransitions}
\usage{
addTransitions(object, transitionList)

\S4method{addTransitions}{Skyline}(object, transitionList)
}
\arguments{
\item{object}{a \code{Skyline} object}

\item{a}{\code{data.frame} of transition information}
}
\description{
addTransitions

Add a transition \code{data.frame} to a \code{Skyline} object. The method writes to
the object in the global environment
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

