% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/calibrate-method.R
\docType{methods}
\name{calibrate}
\alias{calibrate}
\alias{calibrate,skyline-method}
\title{calibrate}
\usage{
calibrate(object, type)

\S4method{calibrate}{skyline}(object, type)
}
\arguments{
\item{a}{\code{skyline} object}
}
\description{
calibrate

Generate optimised calibration curves for each transition
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

