% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveOptimise.R
\name{curveOptimise}
\alias{curveOptimise}
\title{curveOptimise}
\usage{
curveOptimise(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\value{
a \code{data.frame}
}
\description{
Optimise calibration curve
}
\keyword{internal}

