% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/extractCalibrants-method.R
\docType{methods}
\name{extractCalibrants}
\alias{extractCalibrants}
\alias{extractCalibrants,skyline-method}
\title{extractCalibrants}
\usage{
extractCalibrants(object)

\S4method{extractCalibrants}{skyline}(object)
}
\arguments{
\item{a}{\code{skyline} object}
}
\description{
extractCalibrants

Extract all peak data for calibrants and store in \code{calibrants} slot
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

