% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/extractIS.R
\docType{methods}
\name{extractIS}
\alias{extractIS}
\alias{extractIS,skyline-method}
\title{extractIS}
\usage{
extractIS(object)

\S4method{extractIS}{skyline}(object)
}
\arguments{
\item{a}{\code{skyline} object}
}
\description{
extractIS

extract peak data (area, noise and Rt) for the internal standard (IS) from every sample.
Once IS peak data has been added to the \code{internalstd} slot in a \code{skyline} object; it is removed
from the peak tables in the \code{peakInfo} slot.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

