% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCurve.R
\name{fitCurve}
\alias{fitCurve}
\title{Fit equation to calibration curve}
\usage{
fitCurve(x, y, type)
}
\arguments{
\item{x}{a numeric vector of the predictor variable (concentration)}

\item{y}{a numeric vector of the response variable (peak area)}

\item{type}{a character string indicating which equation to use; \code{Lin} = Linear. \code{Quad} = Quadratic,
\code{LogLog} = Logarithimic and \code{Quad-LogLog} = Quadratic-Logarithmic.}
}
\value{
a \code{lm} model object
}
\description{
Select an equation type to use during calibration and subsequent quantificatin
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

