% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internalSTD.R
\docType{methods}
\name{get_internalStd,skyline-method}
\alias{get_internalStd,skyline-method}
\title{get_internalStd}
\usage{
\S4method{get_internalStd}{skyline}(object)
}
\description{
extract peak data (area, noise and Rt) for the internal standard (IS) from every sample.
Once IS peak data has been added to the \code{internalstd} slot in a \code{skyline} object; it is removed
from the peak tables in the \code{peakInfo} slot.
}
