% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/get_peakInfo-method.R
\docType{methods}
\name{get_peakInfo}
\alias{get_peakInfo}
\alias{get_peakInfo,skyline-method}
\title{get_peakInfo}
\usage{
get_peakInfo(object)

\S4method{get_peakInfo}{skyline}(object)
}
\arguments{
\item{object}{a \code{skyline} object}
}
\description{
get_peakInfo

get_peakInfo
}
