% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/integrate-method.R
\docType{methods}
\name{integrate}
\alias{integrate}
\alias{integrate,skyline-method}
\title{integrate}
\usage{
integrate(object)

\S4method{integrate}{skyline}(object)
}
\arguments{
\item{a}{\code{skyline} object}
}
\description{
integrate

integrate
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

