% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{lmeqn}
\alias{lmeqn}
\title{Linear model equation}
\usage{
lmeqn(x, y)
}
\arguments{
\item{x}{numeric vector of \code{x} variable}

\item{u}{numeric vector of \code{y} variable}

\item{NULL}{}
}
\description{
Linear modelling equation for plot annotations
}
\keyword{internal}

