% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotCalibrant-method.R
\docType{methods}
\name{plotCalibrant}
\alias{plotCalibrant}
\alias{plotCalibrant,skyline-method}
\title{plotCalibrant}
\usage{
plotCalibrant(object, name, type)

\S4method{plotCalibrant}{skyline}(object, name, type)
}
\arguments{
\item{a}{\code{skyline} object}
}
\description{
plotCalibrant

Generate optimised calibration curves for each transition
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

