% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{predictModel}
\alias{predictModel}
\title{Predict values}
\usage{
predictModel(m, y)
}
\arguments{
\item{m}{a \code{lm} model object. \seealso{\link{fitCurve}}}

\item{y}{a numeric vector of the response variable (peak area)}
}
\value{
a numeric vector of predicted values
}
\description{
Predict concentration values using response variable and a calibration model
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}

