% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/quantify-method.R
\docType{methods}
\name{quantify}
\alias{quantify}
\alias{quantify,skyline-method}
\title{quantify}
\usage{
quantify(object, area)

\S4method{quantify}{skyline}(object, area)
}
\arguments{
\item{area}{a character string of either \code{ratio} or \code{raw}. If \code{ratio} is selected then peak areas are expressed
as a ratio to the internal stanard. If \code{raw} is selected then the raw peak area is used.}

\item{a}{\code{skyline} object}
}
\description{
quantify

Calculate quantitation results for all non-calibrant samples
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

