% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{skyline_runner}
\alias{skyline_runner}
\title{Skyline CommandLine Runner}
\usage{
skyline_runner(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Skyline CommandLine Runner
}
\keyword{internal}
