% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{validate_phenoData}
\alias{validate_phenoData}
\title{Validate phenoData}
\usage{
validate_phenoData(x, object)
}
\arguments{
\item{x}{a \code{data.frame} of \code{fileName} and \code{sampleName}}

\item{object}{a \code{skyline} object containing filepaths}
}
\description{
Validate phenoData
}
\keyword{internal}
