% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{validate_transitions}
\alias{validate_transitions}
\title{Validate transitions}
\usage{
validate_transitions(x)
}
\arguments{
\item{x}{a \code{data.frame} of Skyline formatted transition data}
}
\description{
Validate transitions
}
\keyword{internal}
