% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/add_filepaths-method.R
\docType{methods}
\name{add_filepaths}
\alias{add_filepaths}
\alias{add_filepaths,skyline-method}
\title{add_filepaths}
\usage{
add_filepaths(object, filepaths)

\S4method{add_filepaths}{skyline}(object, filepaths)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{filepaths}{a vector of valid file (`.raw`, `.mzML` or `.mxXML`) locations for analysis}
}
\description{
add_filepaths

Add the file paths of all raw data to be analysed to a \code{skyline} object. The method writes to
the object in the global environment
}
