% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/addTransitions-method.R
\docType{methods}
\name{addTransitions}
\alias{addTransitions}
\alias{addTransitions,skyline-method}
\title{addTransitions}
\usage{
addTransitions(object, transitionList)

\S4method{addTransitions}{skyline}(object, transitionList)
}
\arguments{
\item{object}{a \code{skyline} object}

\item{a}{\code{data.frame} of transition information. The \code{data.frame} must contain the following columns;
PrecursorName, PrecursorRT, PrecursorMz, ProductMz,PrecursorCharge and ProductCharge}
}
\description{
addTransitions

Add a transition \code{data.frame} to a \code{skyline} object. The method writes to
the object in the global environment
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

