% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/calibrate-method.R
\docType{methods}
\name{calibrate}
\alias{calibrate}
\alias{calibrate,skyline-method}
\title{calibrate}
\usage{
calibrate(object, type, area)

\S4method{calibrate}{skyline}(object, type, area)
}
\arguments{
\item{type}{a character string indicating which regression function to use. The options are;
\itemize{
\item{Lin} {linear}
 \item{LogLog} {logarithmic}
 \item{Quad} {quadratic}
 \item {Quad-LogLog} {quadratic-logarithmic}
}}

\item{area}{a character string of either \code{ratio} or \code{raw}. If \code{ratio} is selected then peak areas are expressed
as a ratio to the internal stanard. If \code{raw} is selected then the raw peak area is used.}

\item{a}{\code{skyline} object}
}
\description{
calibrate

Generate optimised calibration curves for each transition
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

