% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/quantify-method.R
\docType{methods}
\name{quantify}
\alias{quantify}
\alias{quantify,skyline-method}
\title{quantify}
\usage{
quantify(object)

\S4method{quantify}{skyline}(object)
}
\arguments{
\item{a}{\code{skyline} object}
}
\description{
quantify

Calculate quantitation results for all non-calibrant samples
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

