% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{skyline-class}
\alias{skyline-class}
\title{Skyline}
\description{
An \code{S4} class for performing quantitative analysis of SRM-MS data using \code{skyline}
}
\section{Slots}{

\describe{
\item{\code{env}}{class environment}

\item{\code{SkylinePath}}{character of path to \code{SkylineRunner.exe}}

\item{\code{SkylineTransition}}{character of path to Skyline transition file (\code{.sky})}

\item{\code{path}}{character of path where temporary files are written to}

\item{\code{phenoData}}{a runinfo \code{data.frame}}

\item{\code{transitions}}{a \code{data.frame} of transition information. The transitions must be identical to those in the \code{SkyTransitionFile.sky}}

\item{\code{filepaths}}{a character vector of \code{.raw} filepaths}

\item{\code{peakInfo}}{a list of peak picking results}
}}

